fix: block timer
