+++ website/frontend/src/components/pages/Programs/children/BlocksList/BlocksList.scss
@@ -1,7 +1,7 @@
 @import '../../../../../assets/scss/variables';
 
 .block-list {
-  padding-bottom: 4px;
+  padding-bottom: 20px;
 }
 
 .block-list__header {
